Add files via upload
This is a Prototype of Twitter using C++. Initially, we need to sign up to create an account then login into the account. Once the login is successful, we can create tweets and view them, also we can create multiples account with different users names for our friends and get connected. No, Designs were included because I want to make this project only based on CPP.